chore(gatsby): migrate sort to TypeScript (#26369)
* chore(gatsby): migrate sort to TypeScript

* chore(gatsby): remove sort.js
